micro_analyst_full_with_scripts/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
python-dotenv
pydantic
loguru
tenacity
google-generativeai
beautifulsoup4

